refactor(admin): extract the operator surface into an admin blueprint
