+++ tests/test_lea_reports.py
@@ -119,16 +119,16 @@ class TestAdminLeaReports:
 
 
 class TestLeaReportRouteExists:
-    """Static test: verify routes exist in app.py source."""
+    """Static test: verify routes exist in admin.py source."""
 
     def test_lea_report_post_route_in_source(self):
-        with open(os.path.join(PROJECT_ROOT, "app.py")) as f:
+        with open(os.path.join(PROJECT_ROOT, "admin.py")) as f:
             content = f.read()
         assert "/api/internal/lea-report" in content
         assert "X-Internal-Token" in content
 
     def test_admin_lea_reports_route_in_source(self):
-        with open(os.path.join(PROJECT_ROOT, "app.py")) as f:
+        with open(os.path.join(PROJECT_ROOT, "admin.py")) as f:
             content = f.read()
         assert "/admin/lea-reports" in content
         assert "get_lea_reports" in content
